+++ src/main/java/calculator/domain/Operator.java
@@ -3,25 +3,25 @@ package calculator.domain;
 public enum Operator {
     PLUS("+"){
         @Override
-        public int apply(final int operand1, final int operand2){
+        public double apply(final double operand1, final double operand2){
             return operand1+operand2;
         }
     },
     MINUS("-"){
         @Override
-        public int apply(final int operand1, final int operand2){
+        public double apply(final double operand1, final double operand2){
             return operand1-operand2;
         }
     },
     MUL("*"){
         @Override
-        public int apply(final int operand1, final int operand2){
+        public double apply(final double operand1, final double operand2){
             return operand1*operand2;
         }
     },
     DIV("/"){
         @Override
-        public int apply(final int operand1, final int operand2){
+        public double apply(final double operand1, final double operand2){
             if(operand2==0){
                 throw new IllegalArgumentException("0으로 나눌수 없습니다.");
             }
@@ -40,5 +40,5 @@ public enum Operator {
     }
 
 
-    public abstract int apply(int operand1, int operand2);
+    public abstract double apply(double operand1, double operand2);
 }
